Incorporated Dote::SyntaxPass into dote.rb
- Absorbed SyntaxPass into Dote module and covered with a new test in dote_test.rb.
- Deleted SyntaxPass file and tests
