+++ lib/dote.rb
@@ -1,6 +1,5 @@
 require 'vert'
 require_relative 'dote/token_pass'
-require_relative 'dote/syntax_pass'
 require_relative 'dote/code_gen'
 
 module Dote
@@ -22,7 +21,7 @@ module Dote
     if validate_json?(program)
       token_sequence = TokenPass.tokenize_program(program, grammar)
                        .verify_special_forms
-      tree = SyntaxPass.build_tree(token_sequence, grammar)
+      tree = build_tree(token_sequence, grammar)
     else
       validation_pass(program)
     end
@@ -38,6 +37,17 @@ module Dote
     when Regexp.new(MALFORMED_PROGRAM)
       raise SyntaxError, validate_json(program, options)
     end
+  end
+
+  # @param token_seq [TokenSeq]
+  # @return [AbstractSyntaxTree] ParseTree for token_seq
+  def build_tree(token_seq, grammar)
+    parse_tree = grammar.top_rule.parse(token_seq, grammar)[:tree]
+    grammar.eval_tree_attributes(parse_tree)
+  end
+
+  def build_ast(tree, grammar)
+    grammar.convert_to_ast(tree)
   end
 
   def semantic_pass(tree, grammar)
