+++ test/dote_test.rb
@@ -6,7 +6,7 @@ require_relative './test_helpers.rb'
 describe Dote do
 
   include TestHelpers
-  
+
   before do
     @empty_program = get_empty_program
     @unknown_special_form_program = get_unknown_special_form_program
@@ -14,24 +14,23 @@ describe Dote do
     @valid_program = get_tokenizer_sample_program
   end
 
-  describe "empty program" do
-    it ".compile" do
+  describe "#compile" do
+
+    it "valid_program" do
+      Dote.compile(@valid_program).must_be_instance_of Parser::ParseTree
+    end
+
+    it "empty_program" do
       Dote.compile(@empty_program).must_be_nil
     end
-  end
 
-  describe "malformed program" do
-    it ".compile" do
+    it "malformed_program" do
       proc {Dote.compile(@malformed_program)}.must_raise Dote::SyntaxError
     end
-  end
 
-  describe "program_with_unknown_special_forms" do
-    it ".compile" do
+    it "program_with_unknown_special_forms" do
       proc {Dote.compile(@unknown_special_form_program)}
         .must_raise Dote::TokenPass::TokenSeq::UnknownSpecialForm
     end
   end
 end
-
-
